move functions from abstractmake to uploadtrait
