+++ src/services/AbstractMake.php
@@ -10,7 +10,6 @@ declare(strict_types=1);
 
 namespace Elabftw\Services;
 
-use function bin2hex;
 use function dirname;
 use Elabftw\Elabftw\Db;
 use Elabftw\Elabftw\Tools;
@@ -18,8 +17,6 @@ use Elabftw\Exceptions\FilesystemErrorException;
 use Elabftw\Models\AbstractEntity;
 use Elabftw\Traits\UploadTrait;
 use function file_get_contents;
-use function hash;
-use function random_bytes;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
@@ -47,35 +44,6 @@ abstract class AbstractMake
      * The filename for what we are making
      */
     abstract public function getFileName(): string;
-
-    /**
-     * Create a unique long filename with a folder
-     *
-     * @return string the path for storing the file
-     */
-    protected function getLongName(): string
-    {
-        $hash = hash('sha512', bin2hex(random_bytes(16)));
-        $folder = substr($hash, 0, 2);
-
-        return $folder . '/' . $hash;
-    }
-
-    /**
-     * Get the temporary files folder absolute path
-     * Create the folder if it doesn't exist
-     *
-     * @return string absolute path
-     */
-    protected function getTmpPath(): string
-    {
-        $tmpPath = dirname(__DIR__, 2) . '/cache/elab/';
-        if (!is_dir($tmpPath) && !mkdir($tmpPath, 0700, true) && !is_dir($tmpPath)) {
-            throw new FilesystemErrorException("Unable to create the cache directory ($tmpPath)");
-        }
-
-        return $tmpPath;
-    }
 
     /**
      * Get the contents of assets/pdf.min.css
